폴더명 유효성 검증 추가 + 폴더 삭제 시 유저 검증 추가

+++ src/models/folder.js
@@ -1,5 +1,5 @@
 const pool = require('../config/db');
-const { NotFound } = require('../utils/errors');
+const { NotFound, BadRequest } = require('../utils/errors');
 const { ErrorMessage } = require('../utils/response');
 
 module.exports = {
@@ -111,14 +111,16 @@ module.exports = {
     return true;
   },
   deleteFolder: async function (req) {
+    const userId = Number(req.decoded);
     const folderId = Number(req.params.folder_id);
 
-    const sqlDelete = `DELETE FROM folders WHERE folder_id = ?`;
+    const sqlDelete = `DELETE FROM folders WHERE folder_id = ? AND user_id = ?`;
+    const params = [folderId, userId];
 
     const connection = await pool.connection(async (conn) => conn);
     await connection.beginTransaction();
     const [rows] = await connection
-      .query(sqlDelete, folderId)
+      .query(sqlDelete, params)
       .then(await connection.commit())
       .catch(await connection.rollback());
     connection.release();
@@ -127,5 +129,23 @@ module.exports = {
       throw new NotFound(ErrorMessage.folderDeleteError);
     }
     return true;
+  },
+  vaildateFolder: async function (req) {
+    const userId = Number(req.decoded);
+    const folderName = req.body.folder_name;
+
+    const sqlSelect =
+      'SELECT folder_name FROM folders WHERE user_id = ? AND folder_name = ?';
+    const params = [userId, folderName];
+
+    const connection = await pool.connection(async (conn) => conn);
+    const [rows] = await connection.query(sqlSelect, params);
+    connection.release();
+
+    if (rows.length >= 1) {
+      throw new BadRequest(ErrorMessage.vaildateFolder);
+    }
+
+    return false;
   },
 };
